fix get banner

+++ kiosk-solution.Data/ViewModels/UpdateKioskViewModel.cs
@@ -11,8 +11,8 @@ namespace kiosk_solution.Data.ViewModels
     {
         public Guid Id { get; set; }
         public string Name { get; set; }
-        public long Longtitude { get; set; }
-        public long Latitude { get; set; }
+        public double Longtitude { get; set; }
+        public double Latitude { get; set; }
         public Guid KioskLocationId { get; set; }
     }
 }
